social_news/requirements.txt
astroid
attrs
beautifulsoup4
bs4
click
coverage
dill
Flask
Flask-Cors
iniconfig
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
packaging
platformdirs
pluggy
psycopg2
psycopg2-binary
pylint
pytest
pytest-cov
python-dotenv
six
soupsieve
tomlkit
Werkzeug
wrapt